+++ src/app/(auth-required)/compare/page.tsx
@@ -1,9 +1,7 @@
 import { Metadata } from 'next';
 import { ArriveSoon } from '@/components';
 
-export const metadata: Metadata = {
-  title: '통계 비교',
-};
+export const metadata: Metadata = { title: '통계 비교' };
 
 export default function Page() {
   return <ArriveSoon />;
