+++ src/app/(auth-required)/components/header/Section.tsx
