+++ src/app/(auth-required)/leaderboards/components/index.ts
