+++ src/app/(auth-required)/main/components/Section/Graph.tsx
