+++ src/app/(auth-required)/main/components/Section/index.tsx
