+++ src/app/(auth-required)/main/components/Summary/BarContent.tsx
