+++ src/app/(auth-required)/main/components/Summary/SidebarContent.tsx
