+++ src/app/(auth-required)/main/components/index.ts
