+++ src/app/(auth-required)/main/page.tsx
