+++ src/app/(login)/Content.tsx
@@ -1,12 +1,12 @@
 'use client';
 
+import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 import Image from 'next/image';
-import { useMutation } from '@tanstack/react-query';
 import { Input, Button } from '@/components';
-import { LoginVo } from '@/types';
 import { login, sampleLogin } from '@/apis';
+import { LoginVo } from '@/types';
 
 const responsiveStyle =
   "flex items-center gap-5 max-MBI:before:inline-block max-MBI:before:bg-[url('/favicon.png')] max-MBI:before:[background-size:_100%_100%] max-MBI:before:w-16 max-MBI:before:h-16";
