+++ src/app/(login)/page.tsx
@@ -1,10 +1,7 @@
 import { Metadata } from 'next';
-
 import { Content } from './Content';
 
-export const metadata: Metadata = {
-  title: '로그인',
-};
+export const metadata: Metadata = { title: '로그인' };
 
 export default function Page() {
   return <Content />;
