--- src/app/(with-tracker)/layout.tsx
@@ -1,5 +1,0 @@
-import { Suspense } from 'react';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <Suspense>{children}</Suspense>;
-}
