--- src/components/auth-required/index.ts
@@ -1,3 +1,0 @@
-export * from './/header';
-export * from './main';
-export * from './leaderboards';
