--- src/components/common/index.ts
@@ -1,8 +1,0 @@
-export * from './ChannelTalkProvider';
-export * from './QueryProvider';
-export * from './ArriveSoon';
-export * from './Dropdown';
-export * from './Button';
-export * from './Input';
-export * from './Check';
-export * from './Icon';
